remove getIntialProps from article-series and books
